Removed activities and java files for eventGrid, Registration, and User. Refactored Main activity and java files so eventGrid is now the mainactivity
