--- app/src/main/res/layout/activity_event_grid.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/black"
-    tools:context="eventtrackingapp.example.eventtrackingapp.EventGrid">
-
-    <androidx.appcompat.widget.Toolbar
-        android:layout_width="match_parent"
-        android:layout_height="120dp"
-        android:minHeight="?actionBarSize"
-        android:theme="?actionBarTheme"
-        android:background="@color/black">
-        <com.google.android.material.imageview.ShapeableImageView
-            android:layout_width="188dp"
-            android:layout_height="85dp"
-
-
-            android:adjustViewBounds="true"
-            android:scaleType="centerCrop"
-            android:src="@drawable/eventtimelogo"
-            app:layout_constraintBottom_toTopOf="@+id/usernameText"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.454"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:shapeAppearance="@style/CornerCut" />
-
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/actionButtonAdd"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="end"
-
-            android:layout_marginEnd="20dp"
-            android:layout_marginBottom="16dp"
-            android:background="@color/grey"
-            android:backgroundTint="@color/teal_200"
-            android:clickable="true"
-            android:contentDescription="@string/add"
-            android:onClick="onAddClicked"
-            app:borderWidth="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.954"
-            app:layout_constraintStart_toStartOf="parent"
-            app:srcCompat="@android:drawable/ic_input_add"
-            tools:ignore="SpeakableTextPresentCheck,ImageContrastCheck" />
-
-    </androidx.appcompat.widget.Toolbar>
-
-
-
-    <GridView
-        android:id="@+id/eventListView"
-        android:layout_width="match_parent"
-        android:layout_height="fill_parent"
-        android:layout_marginTop="3dp"
-
-        android:dividerHeight="3dp" >
-
-
-    </GridView>
-
-
-
-
-
-</LinearLayout>
