--- app/src/main/res/layout/activity_registration.xml
@@ -1,147 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/black"
-    tools:context="eventtrackingapp.example.eventtrackingapp.Registration">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:orientation="vertical">
-
-        <com.google.android.material.imageview.ShapeableImageView
-            android:id="@+id/shapeableImageView"
-            android:layout_width="214dp"
-            android:layout_height="81dp"
-            android:layout_gravity="center_horizontal"
-
-            android:layout_marginTop="30dp"
-            android:adjustViewBounds="true"
-
-
-            android:fontFamily="serif-monospace"
-            android:scaleType="centerCrop"
-            android:src="@drawable/eventtimelogo"
-            app:shapeAppearance="@style/CornerCut" />
-
-
-        <EditText
-            android:id="@+id/editTextUsername"
-            style="@style/EditTextStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="18dp"
-            android:ems="10"
-            android:hint="@string/editUserName"
-            android:inputType="textPersonName"
-            android:minHeight="48dp"
-            android:layout_gravity="center_horizontal"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.497"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:importantForAutofill="no" />
-
-
-        <EditText
-            android:id="@+id/editTextPassword"
-            style="@style/EditTextStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:ems="10"
-            android:hint="@string/editPassword"
-            android:inputType="textPassword"
-            android:textColor="@color/white"
-            android:layout_gravity="center_horizontal"
-            android:minHeight="48dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.502"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/editTextUsername"
-            android:importantForAutofill="no" />
-
-        <TextView
-            android:id="@+id/permissionsText"
-            style="@style/CardView"
-            android:layout_width="272dp"
-            android:layout_height="73dp"
-            android:layout_marginTop="20dp"
-            android:layout_gravity="center_horizontal"
-            android:background="@drawable/shape"
-            android:fontFamily="monospace"
-            android:text="@string/permissionsText"
-            android:textAlignment="center"
-            android:textSize="19sp"
-            android:textStyle="bold" />
-
-
-        <RadioGroup
-            android:id="@+id/radioGroupYNL"
-            android:layout_width="wrap_content"
-            android:layout_height="144dp"
-            android:layout_marginStart="107dp"
-            android:layout_marginTop="20dp"
-
-            app:layout_constraintTop_toBottomOf="@+id/permissionsText">
-
-            <RadioButton
-                android:id="@+id/radioButtonYes"
-                android:checked='true'
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:fontFamily="monospace"
-
-                android:text="@string/radioButtonYes"
-                android:textColor="@color/orange"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <RadioButton
-                android:id="@+id/radioButtonNo"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-
-                android:text="@string/radioButtonNo"
-                android:textColor="@color/purple_500"
-                android:textSize="18sp"
-                android:textStyle="bold"
-                tools:ignore="TextContrastCheck" />
-
-            <RadioButton
-                android:id="@+id/radioButtonAskLater"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-
-                android:text="@string/radioButtonLater"
-                android:textAlignment="textStart"
-                android:textColor="@color/pink"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-        </RadioGroup>
-
-        <Button
-            android:id="@+id/buttonSaveToDB"
-            style="@style/ButtonStyle"
-            android:layout_width="269dp"
-            android:layout_height="51dp"
-            android:layout_marginBottom="32dp"
-            android:backgroundTint="@color/orange"
-            android:onClick="onSaveClicked"
-            android:text="@string/buttonSave"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginTop="15dp"
-
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.514"
-            app:layout_constraintStart_toStartOf="parent"
-            app:shapeAppearance="@style/CornerCut" />
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
